@@ -4,13 +4,13 @@ build-backend = "hatchling.build"
 
 [project]
 name = "ingest-anything"
-version = "1.1.1"
+version = "1.2.0"
 authors = [
   {name = "Clelia (Astra) Bertelli", email = "astraberte9@gmail.com"}
 ]
 description = "ingest-anything: from data to vector database effortlessly"
 readme = "README.md"
-requires-python = ">=3.10,<3.13"
+requires-python = ">=3.11,<3.13"
 classifiers = [
   "Programming Language :: Python :: 3",
   "License :: OSI Approved :: MIT License",
@@ -20,7 +20,8 @@ dependencies = [
   'chonkie[all]',
   'pydantic',
   'pdfitdown',
-  'llama-index-readers-file'
+  'llama-index-readers-file',
+  'oarc-crawlers'
 ]
 
 [project.urls]
@@ -36,6 +37,8 @@ only-include = ["src/ingest_anything"]
 [tool.hatch.envs.default]
 dependencies = [
   "pytest",
+  "pytest-ordering",
+  "pytest-asyncio",
   "llama-index-vector-stores-milvus",
   "llama-index-vector-stores-qdrant",
   "llama-index-vector-stores-weaviate",
@@ -48,7 +51,7 @@ dependencies = [
 42 = "3.12"
 
 [tool.hatch.envs.default.scripts]
-test = "cp src/ingest_anything/*e*.py tests/ && pytest tests/test_*.py -p no:warnings && rm -rf tests/add_types.py && rm -rf tests/agent_types.py && rm -rf tests/ingestion.py && rm -rf tests/embeddings.py && rm -rf tests/agent.py"
+test = "cp src/ingest_anything/*e*.py tests/ && pytest tests/test_*.py -p no:warnings && rm -rf tests/add_types.py && rm -rf tests/agent_types.py && rm -rf tests/ingestion.py && rm -rf tests/embeddings.py && rm -rf tests/agent.py && rm -rf tests/web_ingestion.py"
 
 [tool.setuptools.packages.find]
 where = ["src"]
