@@ -1,7 +1,17 @@
 try:
-    from .agent_types import IngestAnythingFunctionAgent, IngestAnythingReActAgent, IngestCodeFunctionAgent, IngestCodeReActAgent
+    from .agent_types import (
+        IngestAnythingFunctionAgent,
+        IngestAnythingReActAgent,
+        IngestCodeFunctionAgent,
+        IngestCodeReActAgent,
+    )
 except ImportError:
-    from agent_types import IngestAnythingFunctionAgent, IngestAnythingReActAgent, IngestCodeFunctionAgent, IngestCodeReActAgent
+    from agent_types import (
+        IngestAnythingFunctionAgent,
+        IngestAnythingReActAgent,
+        IngestCodeFunctionAgent,
+        IngestCodeReActAgent,
+    )
 from llama_index.core.llms import LLM
 from llama_index.core.readers.base import BaseReader
 from llama_index.core.vector_stores.types import BasePydanticVectorStore
